+++ src/main/java/com/dlzstudio/iems/renderer/EnergyRelayRenderer.java
@@ -1,0 +1,16 @@
+package com.dlzstudio.iems.renderer;
+
+import com.dlzstudio.iems.IEMSMod;
+import com.dlzstudio.iems.blocks.entity.EnergyRelayBlockEntity;
+import net.minecraft.resources.ResourceLocation;
+import software.bernie.geckolib.renderer.GeoBlockRenderer;
+
+/**
+ * 中继器 GeoLib 渲染器
+ */
+public class EnergyRelayRenderer extends GeoBlockRenderer<EnergyRelayBlockEntity> {
+    
+    public EnergyRelayRenderer() {
+        super(new EnergyRelayModel());
+    }
+}